pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.26.0",
    "python-telegram-bot==20.8",
]
